feat: complete spec sync (#473)
* add project rate limits api

* add usage api

* add usage example

* fix test

* fix test

* admin api keys api

* certificates api

* fix test

* nest admin apis in client.admin()

* nested types for admin apis

* fix example compilation

* types::assistants

* fix imports in the examples

* fix

* fix test

* only unsigned types

* add remanining byot tests

* add realtime apis

* u64 for timestamps

* fix

* upated readme

* updated readme license

* add community project back

* update docs for dynamic dispatch

* update readme

* update readme

* fix test

* updated spec

* cargo fmt

+++ async-openai/src/types/admin/invites.rs
@@ -1,8 +1,7 @@
 use crate::types::OpenAIError;
+use crate::types::OrganizationRole;
 use derive_builder::Builder;
 use serde::{Deserialize, Serialize};
-
-use super::OrganizationRole;
 
 #[derive(Debug, Serialize, Deserialize, Clone, Copy, PartialEq)]
 #[serde(rename_all = "lowercase")]
@@ -54,9 +53,9 @@ pub struct Invite {
     /// `accepted`, `expired`, or `pending`
     pub status: InviteStatus,
     /// The Unix timestamp (in seconds) of when the invite was sent.
-    pub invited_at: u32,
+    pub invited_at: u64,
     /// The Unix timestamp (in seconds) of when the invite expires.
-    pub expires_at: u32,
+    pub expires_at: u64,
     /// The Unix timestamp (in seconds) of when the invite was accepted.
-    pub accepted_at: Option<u32>,
+    pub accepted_at: Option<u64>,
 }
