+++ async-openai/src/types/admin/project_api_keys.rs
@@ -1,6 +1,7 @@
 use serde::{Deserialize, Serialize};
 
-use super::{ProjectServiceAccount, ProjectUser};
+use crate::types::admin::project_service_accounts::ProjectServiceAccount;
+use crate::types::admin::project_users::ProjectUser;
 
 /// Represents an individual API key in a project.
 #[derive(Debug, Serialize, Deserialize)]
@@ -12,7 +13,7 @@ pub struct ProjectApiKey {
     /// The name of the API key.
     pub name: String,
     /// The Unix timestamp (in seconds) of when the API key was created.
-    pub created_at: u32,
+    pub created_at: u64,
     /// The identifier, which can be referenced in API endpoints.
     pub id: String,
     /// The owner of the API key.
